pyproject.toml
[tool.poetry]
name = "thesis"
version = "0.1.0"
description = ""
authors = ["jwkwon0817"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.3.18"
pypdf = "^5.3.0"
langchain-community = "^0.3.17"
python-dotenv = "^1.0.1"
chromadb = "^0.6.3"
langchain-ollama = "^0.2.3"
ollama = "^0.4.7"
langchain-chroma = "^0.2.2"
PyPDF2 = "^3.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
